+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. For commit 
    - shares will check access rules based on user that created the share
 
  **BugFixes**:
+ - passwords with special characters not working properly (#1648)
  - encoded filename navigation issue (#1711)
  - Fix/Improve some behaviors in nextPrevious (#1707)
  - Files recognized as folder if they have the same name as previously deleted folders #1697